+++ src/screens/nc_playlist.h
@@ -19,7 +19,7 @@ typedef struct NcPlaylistScreen {
     int64 start_x;
     int64 width;
     int64 main_start_y;
-    int64 main_height;
+    int32 main_height;
     int64 lines_scrolled;
 
     bool mouse_list_scroll_whole_page;
@@ -63,7 +63,7 @@ NcMenu *nc_playlist_screen_menu(NcPlaylistScreen *screen);
 int64 nc_playlist_screen_height(NcPlaylistScreen *screen);
 void nc_playlist_screen_scroll(NcPlaylistScreen *screen,
                                enum NcScroll where);
-bool nc_playlist_screen_goto_y(NcPlaylistScreen *screen, int64 y);
+bool nc_playlist_screen_goto_y(NcPlaylistScreen *screen, int32 y);
 bool nc_playlist_screen_activate_current(NcPlaylistScreen *screen);
 void nc_playlist_screen_mouse_button_pressed(NcPlaylistScreen *screen,
                                              MEVENT event);
